Move modifiers to own process file
